Update ocean, moving unused ocean stuff to cruft
This drops all usage of DO actions as a guide to items being completed
because DO lies about status.  Also, it a symptom of making cascading
changes and getting a bit out of control.

Tests are just not working, though I did move the action tests to
cruft.  The issues seems to be Mocha in the way.